requirements.txt
anthropic>=0.40.0
playwright>=1.49.0
pandas>=2.2.0
jinja2>=3.1.0
python-dotenv>=1.0.0
